pyproject.toml
[tool.poetry]
name = "python-django"
version = "0.1.0"
description = ""
authors = ["Vitaliy Zhilyaev <jilyaev1987@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
asgiref = "^3.7.2"
gunicorn = "^21.2.0"
attrs = "^23.1.0"
build = "^1.0.3"
cachecontrol = "^0.13.1"
certifi = "^2023.7.22"
charset-normalizer = "^3.3.2"
cleo = "^2.1.0"
colorama = "^0.4.6"
crashtest = "^0.4.1"
distlib = "^0.3.7"
django-debug-toolbar = "^4.2.0"
django-filter = "^23.3"
drf-spectacular = "^0.26.5"
dulwich = "^0.21.6"
fastjsonschema = "^2.18.1"
filelock = "^3.13.1"
idna = "^3.4"
importlib-metadata = "^6.8.0"
inflection = "^0.5.1"
installer = "^0.7.0"
jaraco-classes = "^3.3.0"
jsonschema = "^4.19.2"
jsonschema-specifications = "^2023.7.1"
keyring = "^24.3.0"
more-itertools = "^10.1.0"
msgpack = "^1.0.7"
packaging = "^23.2"
pexpect = "^4.8.0"
pillow = "^10.1.0"
pkginfo = "^1.9.6"
platformdirs = "3.11.0"
pyproject-hooks = "^1.0.0"
pytz = "^2023.3.post1"
poetry = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
